Move some of the macro magic into a separate source file.
